+++ features/admin.applications.v1/pages/applications.tsx
@@ -17,8 +17,6 @@
  */
 
 import { Show } from "@wso2is/access-control";
-import useAuthorization from "@wso2is/admin.authorization.v1/hooks/use-authorization";
-import isLegacyAuthzRuntime from "@wso2is/admin.authorization.v1/utils/get-legacy-authz-runtime";
 import {
     AdvancedSearchWithBasicFilters,
     AppConstants,
@@ -71,7 +69,7 @@ import {
     List,
     PaginationProps
 } from "semantic-ui-react";
-import { useApplicationList, useMyAccountApplicationData, useMyAccountStatus } from "../api";
+import { useApplicationList, useMyAccountApplicationData } from "../api";
 import { useGetApplication } from "../api/use-get-application";
 import { ApplicationList } from "../components/application-list";
 import { MinimalAppCreateWizard } from "../components/wizard/minimal-application-create-wizard";
@@ -122,8 +120,6 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
         [ "data-testid" ]: testId
     } = props;
 
-    const { legacyAuthzRuntime } = useAuthorization();
-
     const { t } = useTranslation();
     const { getLink } = useDocumentation();
 
@@ -153,7 +149,7 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
     const [ filteredApplicationList, setFilteredApplicationList ] = useState<ApplicationListInterface>(null);
     const [ myAccountAppId, setMyAccountAppId ] = useState<string>(null);
 
-    const { isSubOrganization, organizationType } = useGetCurrentOrganizationType();
+    const { organizationType } = useGetCurrentOrganizationType();
 
     const eventPublisher: EventPublisher = EventPublisher.getInstance();
 
@@ -174,15 +170,7 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
         data: myAccountApplication,
         isLoading: isMyAccountApplicationGetRequestLoading,
         error: myAccountApplicationGetRequestError
-    } = useGetApplication(myAccountAppId , !!myAccountAppId && !legacyAuthzRuntime);
-
-    const isSubOrg: boolean = window[ "AppUtils" ].getConfig().organizationName;
-
-    const {
-        data: myAccountStatus,
-        isLoading: isMyAccountStatusLoading,
-        error: myAccountStatusFetchRequestError
-    } = useMyAccountStatus(!isSubOrg && applicationConfig?.advancedConfigurations?.showMyAccountStatus);
+    } = useGetApplication(myAccountAppId , !!myAccountAppId);
 
     /**
      * Set the application id for My Account.
@@ -204,36 +192,20 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
     useEffect(() => {
         let status: boolean = AppConstants.DEFAULT_MY_ACCOUNT_STATUS;
 
-        if (legacyAuthzRuntime) {
-            if (
-                !isApplicationListFetchRequestLoading
-                && !isMyAccountStatusLoading
-                && !isMyAccountApplicationDataFetchRequestLoading
-            ) {
-                if (myAccountStatus) {
-                    const enableProperty: string = myAccountStatus["value"];
-
-                    if (enableProperty && enableProperty === "false") {
-                        status = false;
-                    }
-                }
-            }
-        } else {
-            if (
-                !isApplicationListFetchRequestLoading
+        if (
+            !isApplicationListFetchRequestLoading
                 && !isMyAccountApplicationGetRequestLoading
                 && !isMyAccountApplicationDataFetchRequestLoading
-            ) {
-                if (myAccountApplication) {
-                    status = myAccountApplication?.applicationEnabled;
-                }
+        ) {
+            if (myAccountApplication) {
+                status = myAccountApplication?.applicationEnabled;
             }
         }
+
         setMyAccountStatus(status);
     }, [
         isApplicationListFetchRequestLoading,
         isMyAccountApplicationGetRequestLoading,
-        isMyAccountStatusLoading,
         isMyAccountApplicationDataFetchRequestLoading,
         myAccountAppId
     ]);
@@ -331,40 +303,6 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
     }, [ myAccountApplicationGetRequestError ]);
 
     /**
-     * Handles the application list fetch request error.
-     */
-    useEffect(() => {
-
-        if (!myAccountStatusFetchRequestError) {
-            return;
-        }
-
-        if (
-            myAccountStatusFetchRequestError?.response?.data?.description
-        ) {
-            if (myAccountStatusFetchRequestError.response?.status === 404) {
-                return;
-            }
-            dispatch(addAlert({
-                description: myAccountStatusFetchRequestError.response.data.description ??
-                    t("applications:myaccount.fetchMyAccountStatus.error.description"),
-                level: AlertLevels.ERROR,
-                message: t("applications:myaccount.fetchMyAccountStatus.error.message")
-            }));
-
-            return;
-        }
-
-        dispatch(addAlert({
-            description: t("applications:myaccount.fetchMyAccountStatus" +
-                ".genericError.description"),
-            level: AlertLevels.ERROR,
-            message: t("applications:myaccount.fetchMyAccountStatus" +
-                ".genericError.message")
-        }));
-    }, [ myAccountStatusFetchRequestError ]);
-
-    /**
      * Filter out the system apps from the application list.
      * TODO: This implementation is a temporary one. This filtering should be
      * done from the backend side and should have a seperate endpoint to
@@ -380,13 +318,6 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
                 appList.applications = appList.applications.filter((item: ApplicationListItemInterface) =>
                     !ApplicationManagementConstants.SYSTEM_APPS.includes(item.name)
                     && !ApplicationManagementConstants.DEFAULT_APPS.includes(item.name)
-                );
-            }
-
-            // Remove the M2M application from the application list if the legacy authz runtime is enabled.
-            if (isLegacyAuthzRuntime()) {
-                appList.applications = appList.applications.filter((item: ApplicationListItemInterface) =>
-                    item.templateId !== ApplicationManagementConstants.M2M_APP_TEMPLATE_ID
                 );
             }
 
@@ -513,8 +444,6 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
         if (
             !applicationConfig.advancedConfigurations.showMyAccount
             || UIConfig?.legacyMode?.applicationListSystemApps
-            // Disable the myaccount navigation box for suborgs in legacyauthz runtime.
-            || (isLegacyAuthzRuntime() && isSubOrganization())
         ) {
             return null;
         }
@@ -700,13 +629,9 @@ const ApplicationsPage: FunctionComponent<ApplicationsPageInterface> = (
             data-testid={ `${ testId }-page-layout` }
         >
             {
-                (
-                    legacyAuthzRuntime
-                    || (
-                        !isMyAccountApplicationDataFetchRequestLoading
-                        && myAccountApplicationData?.applications?.length !== 0
-                    )
-                ) && renderTenantedMyAccountLink()
+                !isMyAccountApplicationDataFetchRequestLoading
+                    && myAccountApplicationData?.applications?.length !== 0
+                    && renderTenantedMyAccountLink()
             }
             <ListLayout
                 advancedSearch={ (