+++ features/admin.core.v1/store/reducers/config.ts
@@ -106,6 +106,7 @@ export const commonConfigReducerInitialState: CommonConfigReducerStateInterface<
             dcrConfiguration:"",
             deleteSecret: "",
             deleteSecretType: "",
+            extensions: "",
             externalClaims: "",
             fidoConfigs: "",
             fidoTrustedApps: "",
@@ -268,6 +269,7 @@ export const commonConfigReducerInitialState: CommonConfigReducerStateInterface<
             productVersionConfig: null,
             selfAppIdentifier: "",
             showAppSwitchButton: undefined,
+            showSmsOtpPwdRecoveryFeatureStatusChip: undefined,
             showStatusLabelForNewAuthzRuntimeFeatures: undefined,
             systemAppsIdentifiers: [],
             theme: {