+++ features/admin.server-configurations.v1/constants/index.ts
@@ -18,3 +18,4 @@
 
 export * from "./server-configurations-constants";
 export * from "./governance-connector-constants";
+export * from "./password-recovery-constants";