+++ modules/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wso2is/core
 
+## 2.0.50
+
+### Patch Changes
+
+- [#6478](https://github.com/wso2/identity-apps/pull/6478) [`d6a846aec10597470a11ecce90c27cdf0c6b240d`](https://github.com/wso2/identity-apps/commit/d6a846aec10597470a11ecce90c27cdf0c6b240d) Thanks [@pavinduLakshan](https://github.com/pavinduLakshan)! - Remove legacyAuthzRuntime config from the codebase
+
+## 2.0.49
+
+### Patch Changes
+
+- [#6474](https://github.com/wso2/identity-apps/pull/6474) [`0215c1a74a2f8940d7049e54ec2425b996e50d78`](https://github.com/wso2/identity-apps/commit/0215c1a74a2f8940d7049e54ec2425b996e50d78) Thanks [@ZiyamSanthosh](https://github.com/ZiyamSanthosh)! - Improve logic to generate proper username label when both email as username and multi attribute login features are enabled.
+
 ## 2.0.48
 
 ### Patch Changes